Moved the tests related to issue #1031 into the test_cocotb test.
