--- tests/test_cases/issue_1031/Makefile
@@ -1,8 +1,0 @@
-# Copyright cocotb contributors
-# Licensed under the Revised BSD License, see LICENSE for details.
-# SPDX-License-Identifier: BSD-3-Clause
-
-
-include ../../designs/sample_module/Makefile
-
-MODULE = issue_1031
